--- miniprogram/pages/admin/comment/comment.wxss
@@ -1,3 +1,0 @@
-.margin-top-self {
-    margin-top: 80rpx;
-  }
