+++ miniprogram/pages/admin/index.wxml
@@ -45,12 +45,6 @@
 </view>
 <view class="cu-list menu shadow-lg">
     <view class="cu-item arrow">
-        <view class='content' bindtap='showComment'>
-            <text class='cuIcon-comment text-grey'></text>
-            <text class='text-grey'>评论管理</text>
-        </view>
-    </view>
-    <view class="cu-item arrow">
         <view class='content' bindtap='showArticle'>
             <text class='cuIcon-post text-grey'></text>
             <text class='text-grey'>文章管理</text>
