+++ miniprogram/pages/articleSearch/articleSearch.wxss
@@ -1,0 +1,24 @@
+/* miniprogram/pages/articleSearch/articleSearch.wxss */
+
+/* 文章封面图片容器 */
+.only-img-container {
+  width: 100%;
+  height: 320rpx;
+  border-radius: 6rpx;
+  overflow: hidden;
+  position: relative;
+}
+
+/* 文章封面图片 */
+.only-img {
+  width: 100%;
+  height: 100%;
+  border-radius: 6rpx;
+}
+
+.box {
+  position: sticky;
+  top: 0;
+  z-index: 100;
+}
+
