+++ miniprogram/pages/detail/detail.wxml
@@ -18,8 +18,6 @@
         {{post.totalVisits}}
         <text class="cuIcon-appreciatefill" />
         {{post.totalZans}}
-        <text class="cuIcon-messagefill" />
-        {{post.totalComments}}
       </view>
     </view>
   </view>
@@ -34,59 +32,29 @@
     <ad binderror="adError" bindclose="adClose" unit-id="{{showBannerId}}"></ad>
   </view>
 </view>
-<!-- 评论列表 -->
-<view class="bg-white margin-top-xs">
+
+<!-- 猜你喜欢 -->
+<view class="bg-white margin-top-xs padding-bottom-xl">
   <view class="cu-bar bg-white">
     <view class='action'>
       <text class='cuIcon-titles text-orange '></text>
-      共{{post.totalComments}}条评论
+      猜你喜欢
     </view>
   </view>
-  <!-- 数据不存在 -->
-  <view wx:if="{{nodata}}">
-    <template is="nodata" data="{{ nodata_str }}" />
-  </view>
-  <view class="cu-list menu menu-avatar comment solids-top">
-    <view class="cu-item" wx:for="{{commentList}}" wx:key="idx" wx:for-index="idx" wx:for-item="item" data-idx="{{idx}}">
-      <image class="cu-avatar round" src="{{item.cAvatarUrl}}" mode="aspectFill" data-id="{{item.id}}" data-name="{{item.cNickName}}" data-openid="{{item.cOpenId}}" bindtap="focusComment"></image>
-      <view class='content'>
-        <view class='text-grey' data-id="{{item.id}}" data-name="{{item.cNickName}}" data-openid="{{item.cOpenId}}" bindtap="focusComment">
-          {{item.cNickName}}<view wx:if="{{item.isVip}}" class="cu-tag line-orange radius sm">VIP</view>
-        </view>
-        <view class='text-gray text-content text-df'>{{item.comment}}</view>
-        <view class='padding-sm radius  text-sm text-gray'>
-          <view class="flex" wx:for="{{item.childComment}}" wx:key="id" wx:for-index="id" wx:for-item="childItem">
-            <view data-id="{{item.id}}" data-name="{{childItem.cNickName}}" data-openid="{{childItem.cOpenId}}" bindtap="focusComment">
-              <view class='text-grey'>{{childItem.cNickName}}<view wx:if="{{childItem.isVip}}" class="cu-tag line-orange radius sm">VIP</view>：</view>
-            </view>
-            <view class='flex-sub'>{{childItem.comment}}</view>
-          </view>
-        </view>
-        <view class='flex justify-between'>
-          <view class='text-gray text-df'>{{item.createDate}}</view>
-          <view>
-            <text class="cuIcon-messagefill text-gray margin-left-sm"></text>
-            <text class="text-gray margin-left-xs">{{item.childComment.length}}</text>
-          </view>
-        </view>
+  <view class="cu-list menu">
+    <view class="cu-item" wx:for="{{guessPosts}}" wx:key="id" bindtap="bindPostDetail" id="{{item.id}}">
+      <view class="content" style="width: calc(100% - 30rpx);">
+        <text class="text-blue text-cut" style="text-decoration: underline; display: block;">{{item.title}}</text>
       </view>
     </view>
   </view>
-  <view style="padding:50rpx"></view>
 </view>
+
+<view style="padding:50rpx"></view>
+
 <!-- 底部功能 -->
 <view class="comment-fixed">
-  <!-- <form catchsubmit="formSubmit" report-submit="true"> -->
-  <view class="cu-bar input">
-    <image class="cu-avatar round" src="{{userInfo.avatarUrl}}" mode="aspectFill" bindtap='showMenuBox'></image>
-    <view class='action'>
-      <text class='cuIcon-roundaddfill text-grey' bindtap='showMenuBox'></text>
-    </view>
-    <input class='solid-bottom' maxlength="300" cursor-spacing="10" confirm-type="send" bindinput='commentInput' name="inputComment" value="{{commentContent}}" placeholder="{{placeholder}}" focus="{{focus}}" bindblur="onReplyBlur"></input>
-    <button class='cu-btn bg-green shadow-blur' bindtap="formSubmit">发送</button>
-  </view>
-  <!-- </form> -->
-  <view class="cu-list grid col-4 no-border menu-box {{isShow ? 'emoji-move-in' : 'emoji-move-out'}}">
+  <view class="cu-list grid col-4 no-border menu-box emoji-move-in">
     <view class="cu-item item">
       <view class='cuIcon-share text-olive' style="margin-top:0;"></view>
       <text>转发</text>
@@ -108,38 +76,3 @@
 </view>
 <!-- 弹出层：授权 -->
 <template is="login" data="{{showLogin: showLogin}}"></template>
-
-<!-- 授权弹窗 -->
-<view class="auth-modal" wx:if="{{showAuthModal}}">
-  <view class="auth-modal-mask" bindtap="cancelAuth"></view>
-  <view class="auth-modal-content">
-    <view class="auth-modal-title">完善个人信息</view>
-    <view class="auth-modal-desc">发表评论前，请先设置头像和昵称</view>
-    
-    <view class="auth-form">
-      <view class="auth-form-item auth-avatar-item">
-        <view class="auth-form-label">头像</view>
-        <button class="avatar-wrapper" open-type="chooseAvatar" bindchooseavatar="onChooseAvatar">
-          <image class="avatar-image" src="{{userInfo.tempAvatarUrl || '/images/gravatar.png'}}"></image>
-          <view class="avatar-tips">点击选择</view>
-        </button>
-      </view>
-      
-      <view class="auth-form-item">
-        <view class="auth-form-label">昵称</view>
-        <input 
-          class="auth-input" 
-          type="nickname" 
-          placeholder="请输入昵称" 
-          value="{{userInfo.tempNickName}}"
-          bindinput="onNicknameInput"
-        />
-      </view>
-    </view>
-    
-    <view class="auth-modal-footer">
-      <button class="auth-btn cancel" bindtap="cancelAuth">取消</button>
-      <button class="auth-btn confirm" bindtap="confirmAuth">确认</button>
-    </view>
-  </view>
-</view>
