+++ miniprogram/pages/index/index.wxml
@@ -18,42 +18,15 @@
   </view>
 </view>
 
-<!-- 搜索栏 -->
-<view class='box'>
-  <form bindsubmit='bindconfirm'>
-    <view class="cu-bar search bg-white">
-      <view class='search-form round'>
-        <text class="cuIcon-search"></text>
-        <input type="text" placeholder="搜索感兴趣的文章" value="{{defaultSearchValue}}" bindconfirm="bindconfirm" name="searchContent" confirm-type="search"></input>
-      </view>
-    </view>
-  </form>
-  <scroll-view scroll-x class="bg-white nav solid-bottom">
-    <view class="flex text-center">
-      <view wx:for="{{navItems}}" wx:for-index="idx" data-data="{{item}}" class="cu-item flex-sub {{item.index==tabCur?'text-green cur':''}}" bindtap='tabSelect' data-id="{{item.index}}">
-        {{item.name}}
-      </view>
-    </view>
-  </scroll-view>
-  <scroll-view wx:if="{{showHot}}" scroll-x class="bg-white nav solids-bottom">
-    <view class="padding bg-white">
-      <view wx:for="{{hotItems}}" wx:for-index="idx" data-data="{{item}}" class="cu-tag round lg {{idx==hotCur?'bg-blue':''}}" bindtap='hotSelect' data-id="{{idx}}">
-        {{item}}
-      </view>
-    </view>
-  </scroll-view>
-  <scroll-view wx:if="{{showLabels}}" scroll-x class="bg-white nav solids-bottom">
-    <view class="padding bg-white">
-      <view class="cu-tag round lg {{'全部'==labelCur?'bg-blue':''}}" data-id="全部" bindtap='labelSelect'>全部</view>
-      <view wx:for="{{labelList}}" wx:key="idx" wx:for-index="idx" wx:for-item="item" class="cu-tag round lg {{item.value==labelCur?'bg-blue':''}}" bindtap='labelSelect' data-id="{{item.value}}">
-        {{item.value}}
-      </view>
-    </view>
-  </scroll-view>
-</view>
 <!-- 文章列表 -->
+<view class="cu-bar bg-white margin-top-sm solid-bottom">
+  <view class="action">
+    <text class="cuIcon-titles text-orange"></text>
+    <text class="text-bold">推荐文章</text>
+  </view>
+</view>
 <view class="cu-card dynamic no-card">
-  <view wx:for="{{posts}}" wx:key="idx" wx:for-index="idx" wx:for-item="item" data-idx="{{idx}}" id="{{item.id}}" bindtap="bindPostDetail" class="cu-item shadow solids-bottom">
+  <view wx:for="{{posts}}" wx:key="id" wx:for-index="idx" wx:for-item="item" data-idx="{{idx}}" id="{{item.id}}" bindtap="bindPostDetail" class="cu-item shadow solids-bottom">
     <view class='text-content margin-top-sm'>{{item.title}}</view>
     <view class="grid col-1 flex-sub padding-lr">
       <view class="only-img-container">
@@ -67,8 +40,6 @@
       {{item.totalVisits}}
       <text class="cuIcon-appreciatefill" />
       {{item.totalZans}}
-      <text class="cuIcon-messagefill" />
-      {{item.totalComments}}
     </view>
   </view>
 </view>
